pyproject.toml
[tool.poetry]
name = "budget-app"
version = "0.1.0"
description = "My Budgeting App"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
asyncpg = "^0.30.0"
alembic = "^1.16.4"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.10.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
python-dateutil = "^2.9.0.post0"
httpx = "^0.28.1"
aiofiles = "^24.1.0"
pillow = "^11.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
authlib = "^1.6.0"


[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
black = ">=23.11.0"
isort = ">=5.12.0"
flake8 = ">=6.1.0"
ruff = "^0.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']


[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
ignore = ["F401"]